# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Permify is an open-source authorization service for creating and maintaining fine-grained authorizations accross your individual applications and services.\nPermify converts authorization data as relational tuples into a database you point at. We called that database a Write Database (WriteDB) and it behaves as a centralized data source for your authorization system. You can model of your authorization with Permify's DSL - Permify Schema - and perform access checks with a single API call anywhere on your stack. Access decisions made according to stored relational tuples.",
        "title": "Permify API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/Permify/permify/issues",
            "email": "hello@permify.co"
        },
        "license": {
            "name": "GNU Affero General Public License v3.0"
        },
        "version": "v0.0.0-alpha3"
    },
    "paths": {
        "/permissions/check": {
            "post": {
                "description": "check subject is authorized",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "Permission",
                "operationId": "permissions.check",
                "parameters": [
                    {
                        "description": "check subject is authorized",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/permission.CheckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/permission.CheckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/permissions/expand": {
            "post": {
                "description": "expand relationships according to schema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Permission"
                ],
                "summary": "Permission",
                "operationId": "permissions.expand",
                "parameters": [
                    {
                        "description": "expand relationships according to schema",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/permission.ExpandRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/permission.ExpandResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/relationships/delete": {
            "post": {
                "description": "delete relation tuple",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Relationship"
                ],
                "summary": "Relationship",
                "operationId": "relationships.delete",
                "parameters": [
                    {
                        "description": "delete relation tuple",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/relationship.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tuple.Tuple"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/relationships/read": {
            "post": {
                "description": "read relation tuple(s)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Relationship"
                ],
                "summary": "Relationship",
                "operationId": "relationships.read",
                "parameters": [
                    {
                        "description": "read relation tuple(s)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/relationship.ReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/tuple.Tuple"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/relationships/write": {
            "post": {
                "description": "create new relation tuple",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Relationship"
                ],
                "summary": "Relationship",
                "operationId": "relationships.write",
                "parameters": [
                    {
                        "description": "create new relation tuple",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/relationship.WriteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/tuple.Tuple"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/common.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.HTTPErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {}
            }
        },
...
        "filters.EntityFilter": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "filters.RelationTupleFilter": {
            "type": "object",
            "properties": {
                "entity": {
                    "$ref": "#/definitions/filters.EntityFilter"
                },
                "relation": {
                    "type": "string"
                },
                "subject": {
                    "$ref": "#/definitions/filters.SubjectFilter"
                }
            }
        },
        "filters.SubjectFilter": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "relation": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "permission.CheckRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "depth": {
                    "type": "integer"
                },
                "entity": {
                    "$ref": "#/definitions/tuple.Entity"
                },
                "schema_version": {
                    "type": "string"
                },
                "subject": {
                    "$ref": "#/definitions/tuple.Subject"
                }
            }
        },
        "permission.CheckResponse": {
            "type": "object",
            "properties": {
                "can": {
                    "type": "boolean"
                },
                "decisions": {},
                "remaining_depth": {
                    "type": "integer"
                }
            }
        },
        "permission.ExpandRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "entity": {
                    "$ref": "#/definitions/tuple.Entity"
                },
                "schema_version": {
                    "type": "string"
                }
            }
        },
        "permission.ExpandResponse": {
            "type": "object",
            "properties": {
                "tree": {}
            }
        },
        "relationship.DeleteRequest": {
            "type": "object",
            "properties": {
                "entity": {
                    "$ref": "#/definitions/tuple.Entity"
                },
                "relation": {
                    "type": "string"
                },
                "subject": {
                    "$ref": "#/definitions/tuple.Subject"
                }
            }
        },
        "relationship.ReadRequest": {
            "type": "object",
            "properties": {
                "filter": {
                    "$ref": "#/definitions/filters.RelationTupleFilter"
                }
            }
        },
        "relationship.WriteRequest": {
            "type": "object",
            "properties": {
                "entity": {
                    "$ref": "#/definitions/tuple.Entity"
                },
                "relation": {
                    "type": "string"
                },
                "schema_version": {
                    "type": "string"
                },
                "subject": {
                    "$ref": "#/definitions/tuple.Subject"
                }
            }
        },
...
        "tuple.Entity": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "tuple.Subject": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "relation": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "tuple.Tuple": {
            "type": "object",
            "properties": {
                "entity": {
                    "$ref": "#/definitions/tuple.Entity"
                },
                "relation": {
                    "type": "string"
                },
                "subject": {
                    "$ref": "#/definitions/tuple.Subject"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "api-key-auth",
            "in": "header"
        }
    }
}